% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedFS.R
\name{feature.sel.fun}
\alias{feature.sel.fun}
\title{...}
\usage{
feature.sel.fun(
  x.A,
  y.A,
  x.sam,
  y.sam,
  x.sam2,
  y.sam2,
  method,
  alpha,
  nlambda = 500,
  up.rate = 10,
  drop.rate = 0.5,
  cos.cut
)
}
\arguments{
\item{x.A}{...}

\item{y.A}{...}

\item{x.sam}{...}

\item{y.sam}{...}

\item{x.sam2}{...}

\item{y.sam2}{...}

\item{method}{...}

\item{alpha}{...}

\item{nlambda}{...}

\item{up.rate}{...}

\item{drop.rate}{...}

\item{cos.cut}{...}
}
\description{
...
}
