% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int-feature-selection.R
\name{int.feature.selection}
\alias{int.feature.selection}
\title{Main function for integrative feature selection}
\usage{
int.feature.selection(
  X_full_lst,
  Y_full_lst,
  X_train_lst,
  Y_train_lst,
  X_valid_lst,
  Y_valid_lst,
  lambda_lst = NULL,
  add.ridge = TRUE
)
}
\arguments{
\item{X_full_lst}{...}

\item{Y_full_lst}{...}

\item{X_train_lst}{...}

\item{Y_train_lst}{...}

\item{X_valid_lst}{...}

\item{Y_valid_lst}{...}

\item{lambda_lst}{...}

\item{add.ridge}{...}
}
\description{
Main function for integrative feature selection
}
