% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc-feature-selection.R
\name{loc.feature.selection}
\alias{loc.feature.selection}
\title{Main function for local feature selection}
\usage{
loc.feature.selection(
  X_full,
  Y_full,
  X_train,
  Y_train,
  X_valid,
  Y_valid,
  alpha = 1,
  lambda_lst = NULL,
  up_rate = 10,
  drop_rate = 0.5,
  cos_cut = 0.1,
  add.ridge = TRUE
)
}
\arguments{
\item{X_full}{...}

\item{Y_full}{...}

\item{X_train}{...}

\item{Y_train}{...}

\item{X_valid}{...}

\item{Y_valid}{...}

\item{alpha}{...}

\item{lambda_lst}{...}

\item{up_rate}{...}

\item{drop_rate}{...}

\item{cos_cut}{...}

\item{add.ridge}{...}
}
\description{
Main function for local feature selection
}
